Handle multiline snippets better
Add support for code blocks using markdown syntax. Lines that are
part of the snippet and match the variables regex won't be interpeted as
a variable.

Additionally, add `.cheat.md` as a file extension for cheats so that it
can be rendered on GitHub with proper code formatting and syntax
highlighting.

Fixes #806