Allow cheatsheet items with same descriptions
Prior to the refactor performed in
https://github.com/denisidoro/navi/pull/760, cheatsheet items could have
the same tags and the same description, but a different snippet, and
they would all be listed in the finder.

After the refactor, this behavior has changed and such items will now be
treated as non-unique.

There appears to be no requirement for item descriptions to
be distinct, so this patch restores the earlier behavior.

Fixes: https://github.com/denisidoro/navi/issues/951

Signed-off-by: Erik Gabriel Carrillo <ecarrillo@alumni.utexas.net>